+++ wp-parsely/src/rest-api/settings/class-settings-controller.php
@@ -41,6 +41,7 @@ class Settings_Controller extends REST_API_Controller {
 			new Endpoint_Dashboard_Widget_Settings( $this ),
 			new Endpoint_Editor_Sidebar_Settings( $this ),
 			new Endpoint_Traffic_Boost_Settings( $this ),
+			new Endpoint_Headline_Testing_Settings( $this ),
 		);
 
 		$this->register_endpoints( $endpoints );